complete 70 video part

+++ src/event/events.controller.ts
@@ -1,7 +1,9 @@
 import {
   Body,
+  ClassSerializerInterceptor,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   Logger,
@@ -11,32 +13,31 @@ import {
   Patch,
   Post,
   Query,
+  SerializeOptions,
+  UseGuards,
+  UseInterceptors,
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
 import { CreateEventDto } from './input/create-event.dto';
 import { UpdateEventDto } from './input/update-event.dto';
 import { Event } from './event.entity';
-import { Like, MoreThan, Repository } from 'typeorm';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Attendee } from './attendee.entity';
 import { EventService } from './event.sevice';
 import { ListEvents } from './input/list.events';
+import { User } from 'src/auth/user.entity';
+import { currentUser } from 'src/auth/current-user.decorator';
+import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
 
 @Controller('/events')
+@SerializeOptions({ strategy: 'excludeAll' })
 export class EventsController {
   private readonly logger = new Logger(EventsController.name); // Create an Instance of logger class
 
-  constructor(
-    @InjectRepository(Event)
-    private readonly repositary: Repository<Event>,
-    @InjectRepository(Attendee)
-    private readonly attendeeRepositary: Repository<Attendee>,
+  constructor(private readonly eventService: EventService) {}
 
-    private readonly eventService: EventService,
-  ) {}
   @Get()
   @UsePipes(new ValidationPipe({ transform: true }))
+  @UseInterceptors(ClassSerializerInterceptor)
   async findAll(@Query() filter: ListEvents) {
     this.logger.debug('Hit and run findAll');
     const event =
@@ -50,54 +51,56 @@ export class EventsController {
       );
     return event;
   }
-  @Get('/practice')
-  async practice() {
-    return await this.repositary.find({
-      where: [
-        {
-          id: MoreThan(3),
-          when: MoreThan(new Date('2021-02-12T13:00:00')),
-        },
-        {
-          description: Like('%meet%'),
-        },
-      ],
-      take: 10, //Set to Limit
-      order: {
-        id: 'ASC',
-      },
-    });
-  }
 
-  @Get('/practice2')
-  async practice2() {
-    // return await this.repositary.findOne({
-    //   where: { id: 1 },
-    //   relations: ['attendees'],
-    // });
+  // @Get('/practice')
+  // async practice() {
+  //   return await this.repositary.find({
+  //     where: [
+  //       {
+  //         id: MoreThan(3),
+  //         when: MoreThan(new Date('2021-02-12T13:00:00')),
+  //       },
+  //       {
+  //         description: Like('%meet%'),
+  //       },
+  //     ],
+  //     take: 10, //Set to Limit
+  //     order: {
+  //       id: 'ASC',
+  //     },
+  //   });
+  // }
 
-    const event = await this.repositary.findOne({
-      where: { id: 1 },
-      relations: ['attendees'],
-    });
+  // @Get('/practice2')
+  // async practice2() {
+  //   // return await this.repositary.findOne({
+  //   //   where: { id: 1 },
+  //   //   relations: ['attendees'],
+  //   // });
 
-    // const event = new Event();
-    // event.id = 1;
+  //   const event = await this.repositary.findOne({
+  //     where: { id: 1 },
+  //     relations: ['attendees'],
+  //   });
 
-    const attende = new Attendee();
-    attende.name = 'using cascade function';
-    // attende.event = event;
+  //   // const event = new Event();
+  //   // event.id = 1;
 
-    event.attendees.push(attende);
-    // console.log(add);
+  //   const attende = new Attendee();
+  //   attende.name = 'using cascade function';
+  //   // attende.event = event;
 
-    // await this.attendeeRepositary.save(attende);
-    await this.repositary.save(event);
+  //   event.attendees.push(attende);
+  //   // console.log(add);
 
-    return event;
-  }
+  //   // await this.attendeeRepositary.save(attende);
+  //   await this.repositary.save(event);
+
+  //   return event;
+  // }
 
   @Get(':id')
+  @UseInterceptors(ClassSerializerInterceptor)
   async findOne(@Param('id', ParseIntPipe) id) {
     const event = await this.eventService.getEvent(id);
     if (!event) {
@@ -107,32 +110,47 @@ export class EventsController {
   }
 
   @Post()
-  async create(@Body() input: CreateEventDto) {
-    return await this.repositary.save({
-      ...input,
-      when: new Date(input.when),
-    });
+  @UseGuards(AuthGuardJwt)
+  @UseInterceptors(ClassSerializerInterceptor)
+  async create(@Body() input: CreateEventDto, @currentUser() user: User) {
+    return await this.eventService.createEvent(input, user);
   }
 
   @Patch(':id')
-  async update(@Param('id') id, @Body() input: UpdateEventDto) {
-    const event = await this.repositary.findOne({ where: { id } });
+  @UseGuards(AuthGuardJwt)
+  async update(
+    @Param('id') id,
+    @Body() input: UpdateEventDto,
+    @currentUser() user: User,
+  ) {
+    const event = await this.eventService.getEvent(id);
     if (!event) {
       throw new NotFoundException();
     }
-    return await this.repositary.save({
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when,
-    });
+    if (user.id !== event.organizerId) {
+      throw new ForbiddenException(
+        null,
+        `You ae not authorize to chenge this events`,
+      );
+    }
+
+    return await this.eventService.updateEvent(event, input);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async delete(@Param('id') id) {
-    const result = await this.eventService.deleteEvent(id);
-    if (result.affected !== 1) {
+  async delete(@Param('id') id, @currentUser() user: User) {
+    const event = await this.eventService.getEvent(id);
+    if (!event) {
       throw new NotFoundException();
     }
+    if (user.id !== event.organizerId) {
+      throw new ForbiddenException(
+        null,
+        `You ae not authorize to remove this events`,
+      );
+    }
+    await this.eventService.deleteEvent(id);
   }
 }
